project3/animePlayProject/requirements.txt
args==0.1.0
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.2.6
django-js-asset==2.1.0
django-mptt==0.15.0
gunicorn==21.2.0
packaging==23.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
